Creation of distance_modification_process.cpp
