+++ applications/FluidDynamicsApplication/custom_processes/distance_modification_process.h
@@ -21,7 +21,7 @@
 // External includes
 
 // Project includes
-// #include "includes/define.h"
+#include "includes/define.h"
 #include "processes/process.h"
 #include "includes/cfd_variables.h"
 #include "processes/find_nodal_h_process.h"
@@ -78,40 +78,15 @@ public:
         ModelPart& rModelPart, 
         const bool CheckAtEachStep, 
         const bool NegElemDeactivation,
-        const bool RecoverOriginalDistance)
-        : Process(), mrModelPart(rModelPart) {
+        const bool RecoverOriginalDistance);
 
-        mFactorCoeff = 2.0;
-        mCheckAtEachStep = CheckAtEachStep;
-        mNegElemDeactivation = NegElemDeactivation;
-        mRecoverOriginalDistance = RecoverOriginalDistance;
-    }
-
+    /// Constructor with Kratos parameters.
     DistanceModificationProcess(
         ModelPart& rModelPart,
-        Parameters& rParameters)
-        : Process(), mrModelPart(rModelPart) {
-
-        Parameters default_parameters( R"(
-        {
-            "mesh_id"                                : 0,
-            "model_part_name"                        : "default_model_part_name",
-            "distance_factor"                        : 2.0,
-            "check_at_each_time_step"                : false,
-            "deactivate_full_negative_elements"      : true,
-            "recover_original_distance_at_each_step" : false
-        }  )" );
-
-        rParameters.ValidateAndAssignDefaults(default_parameters);
-
-        mFactorCoeff = rParameters["distance_factor"].GetDouble();
-        mCheckAtEachStep = rParameters["check_at_each_time_step"].GetBool();
-        mNegElemDeactivation = rParameters["deactivate_full_negative_elements"].GetBool();
-        mRecoverOriginalDistance = rParameters["recover_original_distance_at_each_step"].GetBool();
-    }
+        Parameters& rParameters);
 
     /// Destructor.
-    ~DistanceModificationProcess() override{}
+    ~DistanceModificationProcess() override {}
 
     ///@}
     ///@name Operators
